fix "no handler for: …" error from HLS during client detection
The Haskell Language Server gets angry when an unknown LSP request is
sent to it:

    Error detected while processing LspAttach Autocommands for "*":
    LSP[hls] Error condition, please check your setup and/or the [issue tracker](https://github.com/haskell/haskell-language-server/issues):
    LSP: no handler for: "rust-analyzer/analyzerStatus"

It's relatively easy to fix for neovim 0.11+, which provides server_info
in vim.lsp.Client, so we just check the server name, which has been
"rust-analyzer" for the last 5 years [1] and avoid the other heuristics.
On older neovim versions, ferris will still clash with HLS but that's
not an issue for me as I'm on 0.11 already.

[1]: https://github.com/rust-lang/rust-analyzer/blame/eaf37e2c98b66ff7f7a0ac04e4cada39e51fde4b/crates/rust-analyzer/src/bin/main.rs#L277